New plant database backup
Several tables had an ownership specific to me that caused others to
not be able to build the tables. I am not sure how this happened in the
first place, but i fixed it by changing the ownership in pgadmin3. I
double clicked on the owner and it allowed me to change it.