User profile can now be edited
Now we can edit the user profile. Everything but the password and user name can be edited here. The password can be changed via the forgot your password mechanism. This is not sufficient and should be included with the edit profile feature. I believe it is reasonable to require an admin to manually change a username.